touch tracking and tuio messages

+++ OCV_6DoF_Marquer/OCV_6DoF_Marquer.cpp
@@ -331,7 +331,7 @@ void Switchright()
 	{
 		processors[selected_processor]->EnableKeyProcessor(false);
 		selected_processor ++;
-		if(selected_processor >= processors.size()) 
+		if(selected_processor >= (int)processors.size()) 
 			selected_processor = 0;
 		processors[selected_processor]->EnableKeyProcessor();
 	}
